test firld added in bug element

+++ GameTestStudio/src/main/java/com/bartosz/gameteststudio/edit/action/TestEditAction.java
@@ -98,7 +98,7 @@ public class TestEditAction  extends ActionSupport {
 			}
 		}
     	
-    	Test test = TestFabric.getById(Long.parseLong(itemID));
+    	Test test = TestFabric.getById(Integer.parseInt(itemID));
     	
     	platformList = test.getArea().getProject().getPlatformsStringList();
